requirements.txt
# Основные зависимости приложения
Flask==3.1.1
waitress==3.0.2
APScheduler==3.11.0
PyGithub==2.6.1
PyNaCl==1.5.0

# Зависимости для сборки (опционально, можно вынести в requirements-dev.txt)
# nuitka
# zstandard  # Рекомендуется для ускорения Nuitka
